atomic design revision
