+++ src/components/navigations/_HomeStackNavigation/index.js
@@ -1,0 +1,3 @@
+import HomeStackNavigation from './HomeStackNavigation';
+
+export { HomeStackNavigation };
