--- src/components/organism/_Add/index.js
@@ -1,3 +1,0 @@
-import Add from './Add';
-
-export { Add };
