--- src/components/organism/_Home/index.js
@@ -1,3 +1,0 @@
-import Home from './Home';
-
-export { Home };
