--- src/components/organism/_Navigator/index.js
@@ -1,3 +1,0 @@
-import Navigator from './Navigator';
-
-export { Navigator };
