+++ src/components/organisms/_InitalHome/index.js
@@ -1,0 +1,3 @@
+import InitialHome from './InitialHome';
+
+export { InitialHome };
