+++ src/components/organisms/_Scanner/index.js
@@ -1,0 +1,3 @@
+import Scanner from './Scanner';
+
+export { Scanner };
