+++ src/components/organisms/index.js
@@ -1,0 +1,4 @@
+export * from './_Header';
+export * from './_InitalHome';
+export * from './_Scanner';
+export * from './_WifiInput';
